prediction_app/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.8.0
djangorestframework==3.12.4
numpy==1.21.5
pandas==1.3.4
scikit-learn==1.0.1
tensorflow==2.7.0
yfinance==0.1.70